refactor(workflow/service): Remove OpenTelemetry from "workflow" modules
This change removes the OpenTelemetry integration from the OpenDC
Workflow modules. Previously, we chose to integrate OpenTelemetry to
provide a unified way to report metrics to the users.

See the previous commit removing it from the "Compute" modules for the
reasoning behind this change.
